pyproject.toml
[project]
name = "multi-agent-ppm-platform"
version = "0.1.0"
description = "AI-native Multi-Agent Project Portfolio Management Platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Multi-Agent PPM Team"}
]
keywords = ["ppm", "project-management", "portfolio-management", "ai", "agents", "azure"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",

    # Azure SDKs
    "azure-identity>=1.15.0",
    "azure-cosmos>=4.5.1",
    "azure-storage-blob>=12.19.0",
    "azure-servicebus>=7.11.0",
    "azure-eventhub>=5.11.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-monitor-opentelemetry>=1.2.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-exporter-otlp>=1.23.0",
    "cryptography>=42.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",  # PostgreSQL async driver
    "redis>=5.0.0",

    # AI/ML
    "openai>=1.10.0",
    "langchain>=0.1.0",
    "tiktoken>=0.5.2",

    # Utilities
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
    "jsonschema>=4.21.1",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "httpx>=0.26.0",  # For testing API
    "jsonschema>=4.21.1",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/your-org/multi-agent-ppm-platform"
Documentation = "https://github.com/your-org/multi-agent-ppm-platform/tree/main/docs"
Repository = "https://github.com/your-org/multi-agent-ppm-platform"
Issues = "https://github.com/your-org/multi-agent-ppm-platform/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agents", "tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v"
markers = ["asyncio: mark async tests handled by pytest-asyncio or fallback runner"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["fastapi.*", "pydantic.*", "uvicorn.*"]
ignore_missing_imports = true
